% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{plus}
\alias{plus}
\title{Add two numbers together}
\usage{
plus(x, y = 2)
}
\arguments{
\item{x}{a vector of numbers}

\item{y}{another vector of numbers; defaults to number 2.}
}
\value{
the result of adding the two objects together
}
\description{
`plus` returns the sum of the two objects passed in.
}
\examples{
plus(5,3)
plus(4)
}
