% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thing-package.R
\docType{package}
\name{thing-package}
\alias{thing}
\alias{thing-package}
\title{thing: Adds two numbers together}
\description{
See title
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nxskok/thing}
  \item Report bugs at \url{https://github.com/nxskok/thing/issues}
}

}
\author{
\strong{Maintainer}: Ken Butler \email{nxskok@gmail.com}

}
\keyword{internal}
